backend/pyproject.toml
[project]
name = "qyquant-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0,<4.0.0",
    "flask-cors>=4.0.0,<5.0.0",
    "flask-smorest==0.44.0",
    "flask-sqlalchemy==3.1.1",
    "flask-migrate==4.0.7",
    "flask-jwt-extended==4.6.0",
    "python-dotenv==1.0.1",
    "marshmallow==3.21.2",
    "celery==5.3.6",
    "redis==5.0.6",
    "psycopg[binary]==3.1.18",
    "cryptography==42.0.8",
    "psycopg2-binary==2.9.9",
    "jqdatasdk>=1.9.7,<2.0.0",
    "qysp",
]

[tool.uv.sources]
qysp = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
